% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisample_to_single_sample_files.R
\name{convert_cohort_segment_file_to_single_samples}
\alias{convert_cohort_segment_file_to_single_samples}
\title{Convert cohort CNV file to single samples}
\usage{
convert_cohort_segment_file_to_single_samples(
  segment,
  col_sample = "sample",
  col_chromosome = "chromosome",
  col_start = "start",
  col_end = "end",
  col_copynumber = "copyNumber",
  col_minor_cn = "minorAlleleCopyNumber",
  outdir = "cnvs",
  bgzip = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{segment}{a tsv describing somatic copynumber changes by genomic region.}

\item{col_sample}{name of the column containing sample identifiers}

\item{col_chromosome}{column in segment file describing chromosome of the copynumber change}

\item{col_start}{column in segment file describing start position of the copynumber change}

\item{col_end}{column in segment file describing end position of the copynumber change}

\item{col_copynumber}{column in segment file describing the total copynumber}

\item{col_minor_cn}{column in segment file describing the copynumber of the minor allele}

\item{outdir}{Output directory where single-sample segment files will be saved (string).}

\item{bgzip}{should segment files be bgzipped?}

\item{verbose}{verbose mode (flag)}
}
\value{
Invisibly returns NULL. Creates single sample copynumber segment files in the specified output directory.
}
\description{
This function takes a path to a CNV describing segment files (TSV format) and
creates a collection of single-sample copynumber segment files that can be read with
\code{\link[=parse_cnv_to_sigminer]{parse_cnv_to_sigminer()}}. Makes no changes to the data, just splits it up by sample.
}
\examples{
path_copynumber <- system.file(package = "sigstart", "pcawg.3sample.copynumber.segment")
convert_cohort_segment_file_to_single_samples(path_copynumber, outdir = "single_sample_files/cnvs")

}
