% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisample_to_single_sample_files.R
\name{convert_maf_to_vcfs}
\alias{convert_maf_to_vcfs}
\title{Convert MAF-like File to Single Sample VCFs}
\usage{
convert_maf_to_vcfs(
  path,
  outdir = "vcfs",
  col_sample = "Tumor_Sample_Barcode",
  col_chrom = "Chromosome",
  col_start_position = "Start_Position",
  col_end_position = "End_Position",
  col_ref = "Reference_Allele",
  col_alt = "Tumor_Seq_Allele2",
  bgzip = TRUE
)
}
\arguments{
\item{path}{Path to the MAF file (string).}

\item{outdir}{Output directory where VCF files will be saved (string).}

\item{col_sample}{Name of the column describing the sample containing the mutation (string).}

\item{col_chrom}{Name of the column describing the chromosome of the mutation (string).}

\item{col_start_position}{Name of the column describing the start position of the mutation (string).}

\item{col_end_position}{Name of the column describing the end position of the mutation (string).}

\item{col_ref}{Name of the column describing the reference allele (string).}

\item{col_alt}{Name of the column describing the alternate allele (string).}

\item{bgzip}{should segment files be bgzipped?}
}
\value{
Invisibly returns NULL. Creates VCF files in the specified output directory.
}
\description{
This function takes a path to a MAF-like file and creates a collection of single-sample VCFs.
Note that this is \strong{NOT} a lossless operation. Only the most basic chromosome/position/reference/alternate allele information is kept to ensure VCFs are extremely simple.
Also, no filtering is applied to the MAF—every variant in your MAF will be saved in the VCFs.
}
\examples{
path_maf <- system.file(package = "sigstart", "pcawg.3sample.snvs_indel.maf")
convert_maf_to_vcfs(path_maf, outdir = "vcfs")
}
