% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2maf_minimal.R
\name{df2maf_minimal}
\alias{df2maf_minimal}
\title{Convert VCF-like data.frame to MAF}
\usage{
df2maf_minimal(
  data,
  ref_genome,
  keep_all = TRUE,
  col_chrom = "chr",
  col_pos = "pos",
  col_sample_identifier = "sample",
  col_ref = "ref",
  col_alt = "alt",
  missing_to_silent = FALSE,
  col_center = NULL,
  col_entrez_gene_id = NULL,
  col_dbSNP_rsid = NULL,
  col_dbSNP_validation_status = NULL,
  col_matched_normal_sample_identifier = NULL,
  col_sequencer = NULL,
  col_sequence_source = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data.frame with 1 row per mutation in a cohort (data.frame)}

\item{ref_genome}{name of the reference genome used to call variants (string)}

\item{keep_all}{keep all columns in the original data.frame? (flag). If FALSE, only includes the minimal required columns and any column explicitly mapped using \verb{col_<name>} arguments.}

\item{col_chrom}{name of column describing chromosome of the mutation (string)}

\item{col_pos}{name of column describing the 1-based position of the mutation (string)}

\item{col_sample_identifier}{name of column describing the sample containing the mutation (string)}

\item{col_ref}{name of column describing the reference allele (string)}

\item{col_alt}{name of column describing the alternate allele (string)}

\item{missing_to_silent}{Assyne any missing (NA) or empty ('') consequences are 'Silent' mutations}

\item{col_center}{name of column containing the genome sequencing center reporting the variant (string)}

\item{col_entrez_gene_id}{name of column containing entrez gene IDs (string)}

\item{col_dbSNP_rsid}{name of column describing the dbSNP rsid of the variant, or "novel" if there is no dbSNP record (string)}

\item{col_dbSNP_validation_status}{name of column describing the validation status of the variant. Elements must be one of by1000genomes;by2Hit2Allele; byCluster; byFrequency; byHapMap; byOtherPop; bySubmitter; alternate_allele (string)}

\item{col_matched_normal_sample_identifier}{name of column describing the matched normal sample identifier}

\item{col_sequencer}{name of column describing the instrument used to produce data (string)}

\item{col_sequence_source}{name of column describing the molecular assay type used to produce the analytes used for sequencing (string). Elements are usually one of 'WGS', 'WGA', 'WXS', 'RNA-seq', etc}

\item{verbose}{verbose (flag)}
}
\value{
a maf-like data.frame (data.table)
}
\description{
Converts a VEP-like (chr-pos-ref-alt-consequence-sample style) dataframe with to a minimal MAF dataframe.
\strong{Input dataframe should include the columns:}
\itemize{
\item \strong{chr}
\item \strong{pos} (1-based)
\item \strong{sample}
\item \strong{ref}
\item \strong{alt}
\item \strong{consequence} (Sequence Ontology terms)
\item \strong{gene}
}
}
\details{
Alternate column names can be used, but require \strong{col_chrom}, \strong{col_pos}, \strong{col_\if{html}{\out{<etc>}}} arguments to be supplied.
}
\keyword{internal}
