% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_purple_cnv_to_sigminer}
\alias{parse_purple_cnv_to_sigminer}
\title{Parse CNVs to sigminer format}
\usage{
parse_purple_cnv_to_sigminer(
  segment,
  sample_id = "Sample",
  exclude_sex_chromosomes = TRUE
)
}
\arguments{
\item{segment}{path to segment file produced by purple (TUMOR.purple.cnv.somatic.tsv)}

\item{sample_id}{string representing what the sample ID should be. Can be any valid string.}

\item{ignore_gender}{should gen}
}
\value{
data.frame compatible with sigminer (containg the columns sample, chromosome, start, end, segVal, and minor_cn)
}
\description{
Parse CNVs to sigminer format
}
