% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_purple_sv_vcf_to_bedpe}
\alias{parse_purple_sv_vcf_to_bedpe}
\title{Convert Gridss/Purple VCFs to BEDPE format}
\usage{
parse_purple_sv_vcf_to_bedpe(vcf_sv, pass_only = TRUE)
}
\arguments{
\item{vcf_sv}{path to a vcf file produced by GRIDSS / PURPLE (purple SV VCFs are typically the better choice since those are produced post-gripss filtering)}

\item{pass_only}{only return breakpoints where FILTER=PASS}
}
\value{
data.frame of breakpoints in a sigminer compatible format "sample", "chr1", "start1", "end1", "chr2", "start2", "end2", "strand1", "strand2", "svclass"
}
\description{
See \code{\link[sigminer:read_sv_as_rs]{sigminer::read_sv_as_rs()}} for details about the format returned.
Note this sigminer dataset only includes breakpoints (single breakends will be excluded).
By default will only include PASS variants, which will exclude both PON and copy-number inferred breakpoints.
}
\examples{
path_vcf_sv <- system.file("tumor_sample.purple.sv.vcf", package = "sigstart")
parse_purple_sv_vcf_to_bedpe(path_vcf_sv)

}
