% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_tsv_to_sigminer_maf}
\alias{parse_tsv_to_sigminer_maf}
\title{TSV to MAF}
\usage{
parse_tsv_to_sigminer_maf(
  file,
  ref_genome = "uncertain",
  sep = "\\t",
  verbose = FALSE
)
}
\arguments{
\item{file}{path to tsv file describing variants. Must contain columns Sample, Chromosome, Position, Ref, Alt. Position should be 1-based.
If any of these columnames is missing column names will be searched for appropriate alternate names and error if none are found}

\item{ref_genome}{when converting to MAF, what to add as the ref genome field (string)}

\item{sep}{column delimiter (defaults to tab)}

\item{verbose}{verbose mode (flag)}
}
\value{
a maf-like minimal dataframe that can be parsed by \code{\link[sigminer:read_maf]{sigminer::read_maf_minimal()}}
}
\description{
Convert a tabular file to a minimal MAF for signature analysis
}
\examples{

# Define path to TSV file describing mutations
path_tsv <- system.file("chromposrefalt.tsv", package = "sigstart")

# Parse to MAF-like dataframe
maf <- parse_tsv_to_sigminer_maf(path_tsv, verbose = FALSE)

# Inspect MAF
head(maf)

}
