% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_vcf_to_sigminer_maf}
\alias{parse_vcf_to_sigminer_maf}
\title{Convert VCFs into Sigminer-compatible MAF object}
\usage{
parse_vcf_to_sigminer_maf(
  vcf_snv,
  sample_id = NULL,
  include = c("pass", "pass_strict", "all"),
  allow_multisample = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{vcf_snv}{A character string specifying the path to the VCF file.}

\item{sample_id}{string representing the tumour sample identifier.
This is required if you supply a 2-sample tumour normal VCF.
Must be one of the samples described in the VCF.}

\item{include}{Which variants to include in the final data.frame. Valid values include:
\itemize{
\item \strong{pass}: FILTER column is \code{PASS} or \code{.}.
\item \strong{pass_strict}: FILTER column is \code{PASS}.
\item \strong{all}: Include all variants in VCF regardless of FILTER status.
}}

\item{allow_multisample}{Do NOT throw an error for VCFs with >2 samples.
Added since some tumor-normal pipelines add a third sample describing RNA genotypes of mutations found in DNA and these 3-sample VCFs would fail purely because they have 3 samples present.
Whether TRUE/FALSE, this function still assumes every mutation in the VCF relates to the \code{sample_id} supplied, so please exclude any homozygous ref alleles for your sample of interest upstream.}

\item{verbose}{verbose (flag)}
}
\value{
A data.frame with the MAF-like structure, ready for use with Sigminer.
}
\description{
This function reads a VCF file and converts it into a MAF-like structure compatible with Sigminer.
The conversion process includes renaming columns, filtering based on the FILTER field, and standardizing
allele representations.
The VCF file must be either a 1-sample tumor-only VCF or a 2-sample tumor-normal VCF.
If a 2-sample VCF is supplied, you must indicate which is the tumour sample using the\code{sample_id} argument.
Note this function assumes every mutation in the VCF relates to the \code{sample_id} supplied.
(If the VCF includes homozygous ref alleles for your sample of interest please filter these upstream)
}
\examples{
# Convert SNVs and Indels from VCF -> MAF-like structure for Sigminer
path_vcf_snv <- system.file("somatics.vcf", package = "sigstart")
parse_vcf_to_sigminer_maf(path_vcf_snv)
}
