% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_tsv_to_sigminer_maf}
\alias{parse_tsv_to_sigminer_maf}
\title{TSV to MAF}
\usage{
parse_tsv_to_sigminer_maf(
  file,
  col_sample,
  col_chrom,
  col_position,
  col_ref,
  col_alt,
  sep = "\\t"
)
}
\arguments{
\item{file}{path to tsv file describing variants}

\item{col_sample}{name of column containing sample ID}

\item{col_chrom}{name of column containing chromosome}

\item{col_position}{name of column containing position}

\item{col_ref}{name of column describing ref allele}

\item{col_alt}{name of column describing alt allele}

\item{sep}{column delimiter}
}
\value{
a maf-like minimal dataframe that can be parsed by \code{\link[sigminer:read_maf]{sigminer::read_maf_minimal()}}
}
\description{
Convert a tabular file to a minimal MAF for signature analysis
}
