% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_vcf_to_sigminer_maf}
\alias{parse_vcf_to_sigminer_maf}
\title{Convert VCFs into Sigminer-compatible MAF object}
\usage{
parse_vcf_to_sigminer_maf(vcf_snv, pass_only = TRUE)
}
\arguments{
\item{vcf_snv}{A character string specifying the path to the VCF file.}

\item{pass_only}{A logical value indicating whether to filter variants to include only those with a "PASS" filter status. Default is TRUE.}
}
\value{
A data.frame with the MAF-like structure, ready for use with Sigminer.
}
\description{
This function reads a VCF file and converts it into a MAF-like structure compatible with Sigminer.
The conversion process includes renaming columns, filtering based on the FILTER field, and standardizing
allele representations.
}
\examples{
# Convert SNVs and Indels from VCF -> MAF-like structure for Sigminer
path_vcf_snv <- system.file("somatics.vcf", package = "sigstart")
parse_vcf_to_sigminer_maf(path_vcf_snv)
}
