% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstart-package.R
\docType{package}
\name{sigstart-package}
\alias{sigstart}
\alias{sigstart-package}
\title{sigstart: Parse Mutation Files to Formats Compatible with Signature Analysis Toolkits}
\description{
Parse small variants, copy number changes, and structural rearrangements from commmon bioinformatics file formats like VCFs and segment files. Convert these data into formats compatible with signature analysis toolkits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CCICB/sigstart}
  \item Report bugs at \url{https://github.com/CCICB/sigstart/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
